+++ TweetNest/Shared/TweetNestAppDelegate.swift
@@ -24,37 +24,16 @@ class TweetNestAppDelegate: NSObject, ObservableObject {
     let session = Session.shared
     #endif
 
-    @Published private(set) var sessionPersistentContainerStoresLoadingResult: Result<Void, Swift.Error>?
-
     override init() {
         super.init()
 
         Task(priority: .utility) { [self] in
-            await loadSessionPersistentContainerStores()
-
             do {
                 try await session.backgroundTaskScheduler.scheduleBackgroundTasks(for: .active)
 
             } catch {
                 Logger().error("Error occurred while schedule refresh: \(error as NSError, privacy: .public)")
             }
-        }
-    }
-
-    func loadSessionPersistentContainerStores() async {
-        do {
-            try await session.persistentContainer.loadPersistentStores()
-
-            #if DEBUG
-            if TweetNestApp.isPreview {
-                try insertPreviewDataToPersistentContainer()
-            }
-            #endif
-
-            sessionPersistentContainerStoresLoadingResult = .success(Void())
-        } catch {
-            Logger().error("Error occurred while load persistent stores: \(error as NSError, privacy: .public)")
-            sessionPersistentContainerStoresLoadingResult = .failure(error)
         }
     }
 }
