+++ TweetNestKit/Session/Session+Credential.swift
@@ -10,7 +10,7 @@ import CoreData
 import Twitter
 
 extension Session {
-    nonisolated func credential(for accountObjectID: NSManagedObjectID) async throws -> Twitter.Session.Credential? {
+    func credential(for accountObjectID: NSManagedObjectID) async throws -> Twitter.Session.Credential? {
         let context = persistentContainer.newBackgroundContext()
 
         return await context.perform(schedule: .enqueued) {
